+++ CHANGELOG.md
@@ -5,9 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.9] - 2025-04-18
 
-### Added
+### Updated
+- Added style file to package
+
+## [0.2.8] - 2025-04-18
+
+### Updated
 - Added containerWidth prop to MasonryGallery component
 - Added aspectRatio support to GridGallery component
   - Support for global aspectRatio prop